Predefined Models of Quantum States

The submodule QuantumSavory.StatesZoo provides models for many frequently used quantum states.

Autogenerated API list for QuantumSavory.StatesZoo

QuantumSavory.StatesZoo.BarrettKokBellPair — Type

struct BarrettKokBellPair <: QuantumSavory.StatesZoo.AbstractTwoQubitState

Fields:

ηᴬ: Individual channel transmissivity from source A to entanglement swapping station, ∈[0,1]

ηᴮ: Individual channel transmissivity from source B to entanglement swapping station, ∈[0,1]

Pᵈ: Total excess noise (photons per qubit slot) in photon detectors, ≥0, usually ≪1

ηᵈ: Detection efficiency of photon detectors, ∈[0,1]

𝒱: Mode matching parameter for individual interacting photonic pulses with |V| evaluates mode overlap and arg(V) evaluates the carrier phase mismatch, |V|∈[0,1]

m: A single parity bit determined by the click pattern (m = 0 for [0, 1, 1, 0] or [1, 0, 0, 1]; m = 1 for [1, 1, 0, 0] or [0, 0, 1, 1])

metadata

A symbolic representation of the noisy Bell pair state obtained in a Barrett-Kok style protocol (a sequence of two successful entanglement swaps), referred to as the "dual rail photonic qubit swap" in (Prajit Dhara and Guha, 2023) (see eq. C7).

See also BarrettKokBellPairW for the weighted density matrix.

QuantumSavory.StatesZoo.BarrettKokBellPair — Method

BarrettKokBellPair(ηᴬ,ηᴮ)

Asymmetric noiseless perfectly mode-matched case ηᴬ≠ηᴮ, Pᵈ=0, ηᵈ=1, 𝒱=1, m=0

QuantumSavory.StatesZoo.BarrettKokBellPair — Method

BarrettKokBellPair(η)

Symmetric noiseless perfectly mode-matched case ηᴬ=ηᴮ=η, Pᵈ=0, ηᵈ=1, 𝒱=1, m=0

QuantumSavory.StatesZoo.BarrettKokBellPairW — Type

The weighted version of BarrettKokBellPair, i.e. its trace is the probability of successfully heralding a Barrett-Kok Bell pair.

QuantumSavory.StatesZoo.stateexplorer — Function

An interactive explorer for two-qubit states. It returns a new figure.

Requires a Makie plotting backend to be imported.

QuantumSavory.StatesZoo.stateexplorer! — Function

An interactive explorer for two-qubit states. It modifies the given figure.

Requires a Makie plotting backend to be imported.

QuantumSavory.StatesZoo.stateparameters — Function

Return the "interesting" parameters that a state from the StatesZoo has. A constructor that uses only these parameters needs to exist.

Used by stateexplorer to generate the most valuable plots of figures of merit.

QuantumSavory.StatesZoo.stateparametersrange — Function

Return the valid ranges and the "good" value for all parameters listed in stateparameters.