Full API (autogenerated)

Unstable

This is experimental functionality with an unstable API.

QuantumClifford.Experimental.NoisyCircuits.ConditionalGate — Type

A conditional gate that either performs truegate or falsegate, depending on the value of controlbit.

QuantumClifford.Experimental.NoisyCircuits.DecisionGate — Type

A conditional gate that performs one of the gates, depending on the output of decisionfunction applied to the entire classical bit register.

QuantumClifford.Experimental.NoisyCircuits.NoisyBellMeasurement — Type

A Bell measurement in which each of the measured qubits has a chance to have flipped.