Full API

States

Stabilizer states can be represented with the Stabilizer, Destabilizer, MixedStabilizer, and MixedDestabilizer tableau data structures. You probably want to use MixedDestabilizer which supports the widest set of operations.

Moreover, a MixedDestabilizer can be stored inside a Register together with a set of classical bits in which measurement results can be written.

There are convenience constructors for common types of states and operators.

Operations

All of these can be applied on a state with the apply! function. Whether they are deterministic and their computational complexity is listed in the table below. A list of lower-level functions for more control over how an operation is performed is also given.

Type	Deterministic	𝒪(nˣ)	Low-level functions

AbstractOperation

├─ AbstractCliffordOperator

│ ├─ AbstractSymbolicOperator

│ │ ├─ AbstractSingleQubitOperator

│ │ │ ├─ SingleQubitOperator	✔️	n

│ │ │ ├─ sHadamard	✔️	n

│ │ │ ├─ sId1	✔️	n

│ │ │ ├─ sInvPhase	✔️	n

│ │ │ ├─ sPhase	✔️	n

│ │ │ ├─ sX	✔️	n

│ │ │ ├─ sY	✔️	n

│ │ │ └─ sZ	✔️	n

│ │ └─ AbstractTwoQubitOperator

│ │ ├─ sCNOT	✔️	n

│ │ ├─ sCPHASE	✔️	n

│ │ └─ sSWAP	✔️	n

│ │

│ ├─ CliffordOperator	✔️	n³

│ ├─ PauliOperator	✔️	n²

│ └─ SparseGate	✔️	kn²

├─ AbstractMeasurement

│ ├─ PauliMeasurement	❌	n²	project!, projectrand!

│ ├─ sMX	❌	n²	projectX!

│ ├─ sMY	❌	n²	projectY!

│ └─ sMZ	❌	n²	projectZ!

│

└─ Reset	✔️	kn²	reset_qubits!

Autogenerated API list

QuantumClifford.QuantumClifford — Module

A module for using the Stabilizer formalism and simulating Clifford circuits.

QuantumClifford.AbstractSingleQubitOperator — Type

Supertype of all single-qubit symbolic operators.

QuantumClifford.AbstractSymbolicOperator — Type

Supertype of all symbolic operators. Subtype of AbstractCliffordOperator

QuantumClifford.AbstractTwoQubitOperator — Type

Supertype of all two-qubit symbolic operators.

QuantumClifford.CliffordOperator — Type

Clifford Operator specified by the mapping of the basis generators.

julia> CNOT
X_ ⟼ + XX
_X ⟼ + _X
Z_ ⟼ + Z_
_Z ⟼ + ZZ

julia> phase_gate = C"Y
                      Z"
X ⟼ + Y
Z ⟼ + Z

julia> stab = S"XI
                IZ";

julia> entangled = CNOT*stab
+ XX
+ ZZ

You can convert a stabilizer into a Clifford Operator (if necessary, the destabilizers are calculated on the fly):

julia> CliffordOperator(S"Y")
X ⟼ + Z
Z ⟼ + Y

julia> CliffordOperator(S"YY")
ERROR: DimensionMismatch("Input tableau should be square (in which case the destabilizers are calculated) or of size 2n×n (in which case it is used directly).")
[...]

Destabilizer can also be converted (actually, internally, square stabilizer tableaux are first converted to destabilizer tableaux).

julia> d = Destabilizer(S"Y")
+ Z
━━━
+ Y

julia> CliffordOperator(d)
X ⟼ + Z
Z ⟼ + Y

QuantumClifford.Destabilizer — Type

A tableau representation of a pure stabilizer state. The tableau tracks the destabilizers as well, for efficient projections. On initialization there are no checks that the provided state is indeed pure. This enables the use of this data structure for mixed stabilizer state, but a better choice would be to use MixedDestabilizer.

QuantumClifford.MixedDestabilizer — Type

A tableau representation for mixed stabilizer states that keeps track of the destabilizers in order to provide efficient projection operations.

The rank r of the n-qubit tableau is tracked, either so that it can be used to represent a mixed stabilizer state, or so that it can be used to represent an n-r logical-qubit code over n physical qubits. The "logical" operators are tracked as well.

When the constructor is called on an incomplete Stabilizer it automatically calculates the destabilizers and logical operators, following chapter 4 of Daniel Gottesman (1997).

See also: stabilizerview, destabilizerview, logicalxview, logicalzview

QuantumClifford.MixedStabilizer — Type

A slight improvement of the Stabilizer data structure that enables more naturally and completely the treatment of mixed states, in particular when the project! function is used.

QuantumClifford.PauliMeasurement — Type

A Stabilizer measurement on the entirety of the quantum register.

projectrand!(state, pauli) and apply!(state, PauliMeasurement(pauli)) give the same (possibly non-deterministic) result. Particularly useful when acting on Register.

See also: apply!, projectrand!.

QuantumClifford.PauliOperator — Type

A multi-qubit Pauli operator ($±\{1,i\}\{I,Z,X,Y\}^{\otimes n}$).

A Pauli can be constructed with the P custom string macro or by building up one through products and tensor products of smaller operators.

julia> pauli3 = P"-iXYZ"
-iXYZ

julia> pauli4 = 1im * pauli3 ⊗ X
+ XYZX

julia> Z*X
+iY

We use a typical F(2,2) encoding internally. The X and Z bits are stored in a single concatenated padded array of UInt chunks of a bit array.

julia> p = P"-IZXY";

julia> p.xz
2-element Vector{UInt64}:
 0x000000000000000c
 0x000000000000000a

You can access the X and Z bits through getters and setters or through the xview, zview, xbit, and zbit functions.

julia> p = P"XYZ"; p[1]
(true, false)

julia> p[1] = (true, true); p
+ YYZ

QuantumClifford.Register — Type

A register, representing the state of a computer including both a tableaux and an array of classical bits (e.g. for storing measurement results)

QuantumClifford.Reset — Type

Reset the specified qubits to the given state.

QuantumClifford.SingleQubitOperator — Type

A "symbolic" general single-qubit operator which permits faster multiplication than an operator expressed as an explicit tableau.

julia> op = SingleQubitOperator(2, true, true, true, false, true, true) # Tableau components and phases
Symbolic single-qubit gate on qubit 2
X ⟼ - Y
Z ⟼ - X

julia> typeof(op)
SingleQubitOperator

julia> t_op = CliffordOperator(op, 3) # Transforming it back into an explicit tableau representation (specifying the size)
X__ ⟼ + X__
_X_ ⟼ - _Y_
__X ⟼ + __X
Z__ ⟼ + Z__
_Z_ ⟼ - _X_
__Z ⟼ + __Z

julia> typeof(t_op)
CliffordOperator{Vector{UInt8}, Matrix{UInt64}}

julia> CliffordOperator(op, 1, compact=true) # You can also extract just the non-trivial part of the tableau
X ⟼ - Y
Z ⟼ - X

See also: sHadamard, sPhase, sId1, sX, sY, sZ, CliffordOperator

Or simply consult subtypes(QuantumClifford.AbstractSingleQubitOperator) and subtypes(QuantumClifford.AbstractTwoQubitOperator) for a full list. You can think of the s prefix as "symbolic" or "sparse".

QuantumClifford.SparseGate — Type

A Clifford gate, applying the given cliff operator to the qubits at the selected indices.

apply!(state, cliff, indices) and apply!(state, SparseGate(cliff, indices)) give the same result.

QuantumClifford.Stabilizer — Type

Stabilizer, i.e. a list of commuting multi-qubit Hermitian Pauli operators.

Instances can be created with the S custom string macro or as direct sum of other stabilizers.

julia> s = S"XXX
             ZZI
             IZZ"
+ XXX
+ ZZ_
+ _ZZ

julia> s⊗s
+ XXX___
+ ZZ____
+ _ZZ___
+ ___XXX
+ ___ZZ_
+ ____ZZ

It has an indexing API, looking like a list of PauliOperators.

julia> s[2]
+ ZZ_

Pauli operators can act directly on the a stabilizer.

julia> P"YYY" * s
- XXX
+ ZZ_
+ _ZZ

There are a number of ways to create a Stabilizer, including:

generate Stabilizers from a list of Pauli operators

julia> Stabilizer([P"XX", P"ZZ"])
+ XX
+ ZZ

generate Stabilizers from boolean matrices

julia> a = [true true; false false]; b = [false true; true true];

julia> Stabilizer(a, b)
+ XY
+ ZZ

julia> Stabilizer([0x0, 0x2], a, b)
+ XY
- ZZ

initialize an empty Stabilizer and fill it through indexing

julia> s = zero(Stabilizer, 2)
+ __
+ __

julia> s[1,1] = (true, false); s
+ X_
+ __

There are no automatic checks for correctness (i.e. independence of all rows, commutativity of all rows, hermiticity of all rows). The rank (number of rows) is permitted to be less than the number of qubits (number of columns): canonilization, projection, etc. continue working in that case. To great extent this library uses the Stabilizer data structure simply as a tableau. This might be properly abstracted away in future versions.

See also: PauliOperator, canonicalize!

QuantumClifford.Stabilizer — Method

Convert a graph representing a stabilizer state to an explicit Stabilizer.

See also: graphstate

QuantumClifford.sCNOT — Type

A "symbolic" CNOT. See also: AbstractSymbolicOperator

QuantumClifford.sCPHASE — Type

A "symbolic" CPHASE. See also: AbstractSymbolicOperator

QuantumClifford.sHadamard — Type

A "symbolic" single-qubit Hadamard. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.sId1 — Type

A "symbolic" single-qubit Identity operation.

See also: SingleQubitOperator

QuantumClifford.sInvPhase — Type

A "symbolic" single-qubit InvPhase. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.sMX — Type

Symbolic single qubit X measurement. See also Register, projectXrand!, sMY, sMZ

QuantumClifford.sMY — Type

Symbolic single qubit Y measurement. See also Register, projectYrand!, sMX, sMZ

QuantumClifford.sMZ — Type

Symbolic single qubit Z measurement. See also Register, projectZrand!, sMX, sMY

QuantumClifford.sPhase — Type

A "symbolic" single-qubit Phase. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.sSWAP — Type

A "symbolic" SWAP. See also: AbstractSymbolicOperator

QuantumClifford.sX — Type

A "symbolic" single-qubit X. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.sY — Type

A "symbolic" single-qubit Y. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.sZ — Type

A "symbolic" single-qubit Z. See also: SingleQubitOperator, AbstractSymbolicOperator

QuantumClifford.:⊗ — Function

Tensor product between operators or tableaux. See also tensor and tensor_pow.

QuantumClifford.bell — Function

Prepare one or more Bell pairs (with optional phases).

julia> bell()
+ XX
+ ZZ

julia> bell(2)
+ XX__
+ ZZ__
+ __XX
+ __ZZ

julia> bell((true, false))
- XX
+ ZZ

julia> bell([true, false, true, true])
- XX__
+ ZZ__
- __XX
- __ZZ

QuantumClifford.bigram — Method

bigram(state::QuantumClifford.AbstractStabilizer; clip) -> Matrix{Int64}

Get the bigram of a tableau.

It is the list of endpoints of a tableau in the clipped gauge.

If clip=true (the default) the tableau is converted to the clipped gauge in-place before calculating the bigram. Otherwise, the clip gauge conversion is skipped (for cases where the input is already known to be in the correct gauge).

Introduced in Adam Nahum, Jonathan Ruhman, Sagar Vijay, Jeongwan Haah (2017), with a more detailed explanation of the algorithm in Appendix A of Yaodong Li, Xiao Chen, Matthew PA Fisher (2019)

See also: canonicalize_clip!

QuantumClifford.bitview — Function

A view of the classical bits stored with the state

QuantumClifford.canonicalize! — Method

canonicalize!(state::QuantumClifford.AbstractStabilizer; phases, ranks) -> Any

Canonicalize a stabilizer (in place).

Assumes the input is a valid stabilizer (all operators commute and have real phases). It permits redundant generators and identity generators.

julia> ghz = S"XXXX
               ZZII
               IZZI
               IIZZ";

julia> canonicalize!(ghz)
+ XXXX
+ Z__Z
+ _Z_Z
+ __ZZ

julia> canonicalize!(S"XXXX
                       IZZI
                       IIZZ")
+ XXXX
+ _Z_Z
+ __ZZ

Not all rows in the tableau in the next example are independent:

julia> canonicalize!(S"XXXX
                       ZZII
                       IZZI
                       IZIZ
                       IIZZ")
+ XXXX
+ Z__Z
+ _Z_Z
+ __ZZ
+ ____

In cases of lower rank, more advanced tableau structures might be better. For instance the MixedStabilizer or MixedDestabilizer structures (you can read more about them in the Data Structures section of the documentation).

If phases=false is set, the canonicalization does not track the phases in the tableau, leading to significant (constant factor) speedup.

julia> s = S"-ZX
              XZ"
- ZX
+ XZ

julia> canonicalize!(copy(s), phases=false)
- XZ
+ ZX

julia> canonicalize!(copy(s))
+ XZ
- ZX

If ranks=true is set, the last pivot indices for the X and Z stage of the canonicalization are returned as well.

julia> s = S"XXXX
             ZZII
             IZIZ
             ZIIZ";

julia> canonicalize!(s, ranks=true)
(+ XXXX
+ Z__Z
+ _Z_Z
+ ____, 1, 3)

Based on Hector J Garcia, Igor L Markov, Andrew W Cross (2012).

See also: canonicalize_rref!, canonicalize_gott!

QuantumClifford.canonicalize_clip! — Method

canonicalize_clip!(state::QuantumClifford.AbstractStabilizer; phases) -> QuantumClifford.AbstractStabilizer

Fix the clipped gauge of a stabilizer (in place).

Assumes the input is a valid full-rank stabilizer (all operators commute and have real phases).

julia> s = S"- X_ZX_X
             + XXYZ__
             - YZ_Z_X
             - XZX__Y
             + _Z_Y_Y
             - ____Z_";

julia> canonicalize_clip!(s)
- X_XY__
+ YZY___
+ _XZX__
- _ZYX_Z
- __YZ_X
- ____Z_

If phases=false is set, the canonicalization does not track the phases in the tableau, leading to a significant speedup.

Introduced in Adam Nahum, Jonathan Ruhman, Sagar Vijay, Jeongwan Haah (2017), with a more detailed explanation of the algorithm in Appendix A of Yaodong Li, Xiao Chen, Matthew PA Fisher (2019)

See also: canonicalize!, canonicalize_rref!, canonicalize_gott!.

QuantumClifford.canonicalize_gott! — Method

Inplace Gottesman canonicalization of a tableau.

This uses different canonical form from canonicalize!. It is used in the computation of the logical X and Z operators of a MixedDestabilizer.

It returns the (in place) modified state, the indices of the last pivot of both Gaussian elimination steps, and the permutations that have been used to put the X and Z tableaux in standard form.

Based on Daniel Gottesman (1997).

See also: canonicalize!, canonicalize_rref!

QuantumClifford.canonicalize_rref! — Method

canonicalize_rref!(state::QuantumClifford.AbstractStabilizer, colindices; phases) -> Tuple{QuantumClifford.AbstractStabilizer, Any}

Canonicalize a stabilizer (in place) along only some columns.

This uses different canonical form from canonicalize!. It also indexes in reverse in order to make its use in traceout! more efficient. Its use in traceout! is its main application.

It returns the (in place) modified state and the index of the last pivot.

Based on Koenraad MR Audenaert, Martin B Plenio (2005).

See also: canonicalize!, canonicalize_gott!

QuantumClifford.canonicalize_rref! — Method

canonicalize_rref!(state::QuantumClifford.AbstractStabilizer; phases) -> Tuple{QuantumClifford.AbstractStabilizer, Any}

QuantumClifford.clifford_cardinality — Method

The size of the Clifford group over a given number of qubits, possibly modulo the phases.

For n qubits, not accounting for phases is 2ⁿⁿΠⱼ₌₁ⁿ(4ʲ-1). There are 4ⁿ different phase configurations.

See also: enumerate_cliffords.

QuantumClifford.comm — Method

Check whether two operators commute.

0x0 if they commute, 0x1 if they anticommute.

julia> P"XX"*P"ZZ", P"ZZ"*P"XX"
(- YY, - YY)

julia> comm(P"ZZ", P"XX")
0x00

julia> comm(P"IZ", P"XX")
0x01

QuantumClifford.destabilizerview — Method

A view of the subtableau corresponding to the destabilizer. See also tab, stabilizerview, logicalxview, logicalzview

QuantumClifford.entanglement_entropy — Function

Get bipartite entanglement entropy of a subsystem

Defined as entropy of the reduced density matrix.

It can be calculated with multiple different algorithms, the most performant one depending on the particular case.

Currently implemented are the :clip, :graph, and :rref algorithms. Benchmark your particular case to choose the best one.

QuantumClifford.entanglement_entropy — Method

Get bipartite entanglement entropy by first converting the state to a graph and computing the rank of the adjacency matrix.

Based on M. Hein, W. Dür, J. Eisert, R. Raussendorf, M. Van den Nest, H.-J. Briegel (2006).

QuantumClifford.entanglement_entropy — Method

Get bipartite entanglement entropy by converting to RREF form (i.e., partial trace form).

The state will be partially canonicalized in an RREF form.

See also: canonicalize_rref!, traceout!.

QuantumClifford.entanglement_entropy — Method

Get bipartite entanglement entropy of a contiguous subsystem by passing through the clipped gauge.

If clip=false is set the canonicalization step is skipped, useful if the input state is already in the clipped gauge.

See also: bigram, canonicalize_clip!

QuantumClifford.enumerate_cliffords — Method

Give the i-th n-qubit Clifford operation, where i∈{1..2ⁿⁿΠⱼ₌₁ⁿ(4ʲ-1)}

The algorithm is detailed in Robert Koenig, John A Smolin (2014).

See also: symplecticGS, clifford_cardinality.

QuantumClifford.enumerate_cliffords — Method

Give all n-qubit Clifford operations.

The algorithm is detailed in Robert Koenig, John A Smolin (2014).

See also: symplecticGS, clifford_cardinality.

QuantumClifford.enumerate_phases — Method

Given an operator, return all operators that have the same tableau but different phases.

julia> length(collect(enumerate_phases(tCNOT)))
16

See also: enumerate_cliffords, clifford_cardinality.

QuantumClifford.enumerate_phases — Method

Given a set of operators, return all operators that have the same tableaux but different phases.

julia> length(collect(enumerate_phases(enumerate_cliffords(2))))
11520

See also: enumerate_cliffords, clifford_cardinality.

QuantumClifford.enumerate_single_qubit_gates — Method

Generate a symbolic single-qubit gate given its index. Optionally, set non-trivial phases.

See also: enumerate_cliffords.

QuantumClifford.expect — Method

expect(p::PauliOperator, st::AbstractStabilizer)

Compute the expectation value of a Pauli operator p on a stabilizer state st. This function will allocate a temporary copy of the stabilizer state st.

QuantumClifford.generate! — Method

Generate a Pauli operator by using operators from a given the Stabilizer.

It assumes the stabilizer is already canonicalized. It modifies the Pauli operator in place, generating it in reverse, up to a phase. That phase is left in the modified operator, which should be the identity up to a phase. Returns the new operator and the list of indices denoting the elements of stabilizer that were used for the generation.

julia> ghz = S"XXXX
               ZZII
               IZZI
               IIZZ";

julia> canonicalize!(ghz)
+ XXXX
+ Z__Z
+ _Z_Z
+ __ZZ

julia> generate!(P"-ZIZI", ghz)
(- ____, [2, 4])

When the Pauli operator can not be generated by the given tableau, nothing is returned.

julia> generate!(P"XII",canonicalize!(S"ZII")) === nothing
true

julia> generate!(P"XII",canonicalize!(S"XII")) === nothing
false

QuantumClifford.gf2_H_to_G — Method

For a given F(2,2) parity check matrix, return the generator matrix.

QuantumClifford.gf2_gausselim! — Method

Gaussian elimination over the binary field.

QuantumClifford.gf2_invert — Method

Invert a binary matrix.

QuantumClifford.gf2_isinvertible — Method

Check whether a binary matrix is invertible.

QuantumClifford.ghz — Function

Prepare a GHZ state of n qubits.

julia> ghz()
+ XXX
+ ZZ_
+ _ZZ

julia> ghz(2)
+ XX
+ ZZ

julia> ghz(4)
+ XXXX
+ ZZ__
+ _ZZ_
+ __ZZ

QuantumClifford.graph_gate — Method

A helper function converting the gate indices from graphstate into a Clifford operator.

julia> s = S" XXX
              YZ_
             -_ZZ";

julia> graph, h_idx, ip_idx, z_idx = graphstate(s);

julia> gate = graph_gate(h_idx, ip_idx, z_idx, nqubits(s));

julia> apply!(s, gate) # This is now a graph state (notice you need to multiply row 1 by row 2)
+ YYZ
+ XZ_
+ _ZX

julia> canonicalize!(s) == canonicalize!(Stabilizer(graph))
true

See also: graph_gatesequence

QuantumClifford.graph_gatesequence — Method

A helper function converting the gate indices from graphstate into a sequence of gates.

julia> s = S" XXX
              YZ_
             -_ZZ";

julia> graph, h_idx, ip_idx, z_idx = graphstate(s);

julia> gates = graph_gatesequence(h_idx, ip_idx, z_idx);

julia> for gate in vcat(gates...) apply!(s, gate) end

julia> s # This is now a graph state (notice you need to multiply row 1 by row 2)
+ YYZ
+ XZ_
+ _ZX

julia> canonicalize!(s) == canonicalize!(Stabilizer(graph))
true

See also: graph_gatesequence

QuantumClifford.graphstate! — Method

An in-place version of graphstate.

QuantumClifford.graphstate — Method

Convert any stabilizer state to a graph state

Graph states are a special type of entangled stabilizer states that can be represented by a graph. For a graph $G=(V,E)$ the corresponding stabilizers are $S_v = X_v \prod_{u ∈ N(v)} Z_u$. Notice that such tableau rows contain only a single X operator. There is a set of single qubit gates that converts any stabilizer state to a graph state.

This function returns the graph state corresponding to a stabilizer and the gates that might be necessary to convert the stabilizer into a state representable as a graph.

For a tableau stab you can convert it with:

graph, hadamard_idx, iphase_idx, flips_idx = graphstate()

where graph is the graph representation of stab, and the rest specifies the single-qubit gates converting stab to graph: hadamard_idx are the qubits that require a Hadamard gate (mapping X ↔ Z), iphase_idx are (different) qubits that require an inverse Phase gate (Y → X), and flips_idx are the qubits that require a phase flip (Pauli Z gate), after the previous two sets of gates.

julia> using Graphs

julia> s = S" XXX
              ZZ_
             -_ZZ";

julia> g, h_idx, ip_idx, z_idx = graphstate(s);

julia> collect(edges(g))
2-element Vector{Graphs.SimpleGraphs.SimpleEdge{Int64}}:
 Edge 1 => 2
 Edge 1 => 3

julia> h_idx
2-element Vector{Int64}:
 2
 3

julia> ip_idx
Int64[]

julia> z_idx
1-element Vector{Int64}:
 3

The Graphs.jl library provides many graph-theory tools and the MakieGraphs.jl library provides plotting utilies for graphs.

You can directly call the graph constructor on a stabilizer, if you just want the graph and do not care about the Clifford operation necessary to convert an arbitrary state to a state representable as a graph:

julia> collect(edges( Graph(bell()) ))
1-element Vector{Graphs.SimpleGraphs.SimpleEdge{Int64}}:
 Edge 1 => 2

For a version that does not copy the stabilizer, but rather performs transformations in-place, use graphstate!. It would perform canonicalize_gott! on its argument as it finds a way to convert it to a graph state.

QuantumClifford.logdot — Method

Logarithm of the inner product between to Stabilizer states.

If the result is nothing, the dot inner product is zero. Otherwise the inner product is 2^(-logdot/2).

The actual inner product can be computed with LinearAlgebra.dot.

Based on Hector J Garcia, Igor L Markov, Andrew W Cross (2012).

QuantumClifford.logicalxview — Method

A view of the subtableau corresponding to the logical X operators. See also tab, stabilizerview, destabilizerview, logicalzview

QuantumClifford.logicalzview — Method

A view of the subtableau corresponding to the logical Z operators. See also tab, stabilizerview, destabilizerview, logicalxview

QuantumClifford.nqubits — Method

The number of qubits of a given state.

QuantumClifford.prodphase — Method

Get the phase of the product of two Pauli operators.

Phase is encoded as F(4) in the low qubits of an UInt8.

julia> P"ZZZ"*P"XXX"
-iYYY

julia> prodphase(P"ZZZ", P"XXX")
0x03

julia> prodphase(P"XXX", P"ZZZ")
0x01

QuantumClifford.project! — Method

project!(state, pauli::PauliOperator; keep_result, phases) -> Tuple{MixedStabilizer, Any, Any}

Project the state of a Stabilizer on the two eigenspaces of a Pauli operator.

Assumes the input is a valid stabilizer. The projection is done inplace on that stabilizer and it does not modify the projection operator.

It returns

a stabilizer that might not be in canonical form

the index of the row where the non-commuting operator was (that row is now equal to pauli; its phase is not updated and for a faithful measurement simulation it needs to be randomized by the user)

and the result of the projection if there was no non-commuting operator (nothing otherwise)

If keep_result==false that result of the projection in case of anticommutation is not computed, sparing a canonicalization operation. This canonicalization operation is the only one potentially of cubic complexity. The rest of the calculations are of quadratic complexity.

If you need to measure a single qubit instead of a multiqubit Pauli operator, the faster projectX!, projectY!, and projectZ! are available.

For less boilerplate and automatic randomization of the phase use projectrand!.

Here is an example of a projection destroying entanglement:

julia> ghz = S"XXXX
               ZZII
               IZZI
               IIZZ";

julia> canonicalize!(ghz)
+ XXXX
+ Z__Z
+ _Z_Z
+ __ZZ

julia> state, anticom_index, result = project!(ghz, P"ZIII");

julia> state
+ Z___
+ Z__Z
+ _Z_Z
+ __ZZ

julia> canonicalize!(state)
+ Z___
+ _Z__
+ __Z_
+ ___Z

julia> anticom_index, result
(1, nothing)

And an example of projection consistent with the stabilizer state.

julia> s = S"ZII
             IXI
             IIY";

julia> canonicalize!(s)
+ _X_
+ __Y
+ Z__

julia> state, anticom_index, result = project!(s, P"-ZII");

julia> state
+ _X_
+ __Y
+ Z__

julia> anticom_index, result
(0, 0x02)

While not the best choice, Stabilizer can be used for mixed states, simply by providing an incomplete tableau. In that case it is possible to attempt to project on an operator that can not be generated by the provided stabilizer operators. In that case we have anticom_index==rank and result===nothing, where rank is the the new rank of the tableau, one more than the number of rows in the initial tableau. However, if keep_result was set to false, then anticom_index would stay at zero.

julia> s = S"XZI
             IZI";

julia> project!(s, P"IIX")
(+ X__
+ _Z_, 3, nothing)

If we had used MixedStabilizer we would have added the projector to the list of stabilizers.

julia> s = one(MixedStabilizer, 2, 3)
Rank 2 stabilizer
+ Z__
+ _Z_

julia> project!(s, P"IIX")
(Rank 3 stabilizer
+ Z__
+ _Z_
+ __X, 3, nothing)

However, MixedDestabilizer would be an even better choice as it has $\mathcal{O}(n^2)$ complexity instead of the $\mathcal{O}(n^3)$ complexity of *Stabilizer.

julia> s = one(MixedDestabilizer, 2, 3)
Rank 2 stabilizer
+ X__
+ _X_
━━━━━
+ __X
━━━━━
+ Z__
+ _Z_
━━━━━
+ __Z

julia> project!(s, P"IIX")
(Rank 3 stabilizer
+ X__
+ _X_
+ __Z
═════
+ Z__
+ _Z_
+ __X
═════
, 3, nothing)

See the "Datastructure Choice" section in the documentation for more details.

See also: projectX!, projectY!, projectZ!, projectrand!

QuantumClifford.project! — Method

project!(state::MixedStabilizer, pauli::PauliOperator; phases) -> Tuple{MixedStabilizer, Any, Any}

When using project! on MixedStabilizer it automates some of the extra steps we encounter when implicitly using the Stabilizer datastructure to represent mixed states. Namely, it helps when the projector is not among the list of stabilizers:

julia> s = S"XZI
             IZI";

julia> ms = MixedStabilizer(s)
Rank 2 stabilizer
+ X__
+ _Z_

julia> project!(ms, P"IIY")
(Rank 3 stabilizer
+ X__
+ _Z_
+ __Y, 3, nothing)

Similarly to project! on Stabilizer, this function has cubic complexity when the Pauli operator commutes with all rows of the tableau. Most of the time it is better to simply use MixedDestabilizer representation.

Unlike other project! methods, this one does not allow for keep_result=false, as the correct rank or anticommutation index can not be calculated without the expensive (cubic) canonicalization operation required by keep_result=true.

See the "Datastructure Choice" section in the documentation for more details.

See also: projectX!, projectY!, projectZ!.

QuantumClifford.projectX! — Method

Measure a given qubit in the X basis. A faster special-case version of project!.

See also: project!, projectXrand!, projectY!, projectZ!.

QuantumClifford.projectXrand! — Method

projectXrand!(state, qubit) -> Tuple{Register, UInt8}

Project qubit of state along the X axis and randomize the phase if necessary.

Lower boilerplate version of project!.

See also: project!, projectX!, projectZrand!, projectYrand!

QuantumClifford.projectY! — Method

Measure a given qubit in the Y basis. A faster special-case version of project!.

See also: project!, projectYrand!, projectX!, projectZ!.

QuantumClifford.projectYrand! — Method

projectYrand!(state, qubit) -> Tuple{Register, UInt8}

Project qubit of state along the Y axis and randomize the phase if necessary.

Lower boilerplate version of project!.

See also: project!, projectY!, projectXrand!, projectZrand!

QuantumClifford.projectZ! — Method

Measure a given qubit in the Z basis. A faster special-case version of project!.

See also: project!, projectZrand!, projectY!, projectX!.

QuantumClifford.projectZrand! — Method

projectZrand!(state, qubit) -> Tuple{Register, UInt8}

Project qubit of state along the Z axis and randomize the phase if necessary.

Lower boilerplate version of project!.

See also: project!, projectZ!, projectXrand!, projectYrand!

QuantumClifford.projectrand! — Method

projectrand!(state, pauli) -> Tuple{QuantumClifford.AbstractStabilizer, UInt8}

Measure pauli operator on state and randomize the phase if necessary.

Lower boilerplate version of project!.

See also: project!, projectXrand!, projectZrand!, projectYrand!

QuantumClifford.puttableau! — Method

Put source tableau in target tableau at given row and column. Assumes target location is zeroed out.

QuantumClifford.quantumstate — Function

Only the quantum part of the state (excluding classical bits)

QuantumClifford.random_clifford — Method

A random Clifford operator generated by the Bravyi-Maslov Algorithm 2 from Sergey Bravyi, Dmitri Maslov (2021).

QuantumClifford.random_clifford1 — Method

Random symbolic single-qubit Clifford applied to qubit at index qubit.

See also: SingleQubitOperator, random_clifford

QuantumClifford.random_destabilizer — Method

A random Stabilizer/Destabilizer tableau generated by the Bravyi-Maslov Algorithm 2 from Sergey Bravyi, Dmitri Maslov (2021).

QuantumClifford.random_pauli — Method

A random Pauli operator on n qubits.

Use realphase=true to get operators with phase ±1 (excluding ±i). nophase=true sets the phase to +1.

Optionally, a "flip" probability p can be provided specified, in which case each bit is set to I with probability 1-p and to X, Y, or Z, each with probability p. Useful for simulating Pauli noise.

QuantumClifford.random_stabilizer — Method

A random Stabilizer tableau generated by the Bravyi-Maslov Algorithm 2 from Sergey Bravyi, Dmitri Maslov (2021).

QuantumClifford.reset_qubits! — Method

reset_qubits!(s::MixedDestabilizer, newstate::QuantumClifford.AbstractStabilizer, qubits; phases) -> MixedDestabilizer

QuantumClifford.reset_qubits! — Method

reset_qubits!(s::MixedStabilizer, newstate, qubits; phases) -> MixedStabilizer

QuantumClifford.reset_qubits! — Method

reset_qubits!(s::Stabilizer, newstate, qubits; phases) -> Any

Reset a given set of qubits to be in the state newstate. These qubits are traced out first, which could lead to "nonlocal" changes in the tableau.

QuantumClifford.single_x — Method

A multiqubit operator corresponding to all identities except for Pauli X at i. See also: sX, sMX

QuantumClifford.single_y — Method

A multiqubit operator corresponding to all identities except for Pauli Y at i. See also: sY, sMY

QuantumClifford.single_z — Method

A multiqubit operator corresponding to all identities except for Pauli Z at i. See also: sY, sMY

QuantumClifford.stab_to_gf2 — Method

The F(2,2) matrix of a given stabilizer, represented as the concatenation of two binary matrices, one for X and one for Z.

QuantumClifford.stabilizerview — Method

A view of the subtableau corresponding to the stabilizer. See also tab, destabilizerview, logicalxview, logicalzview

QuantumClifford.symplecticGS — Method

Perform the Symplectic Gram-Schmidt procedure that gives a Clifford operator canonically related to a given Pauli operator.

The algorithm is detailed in Robert Koenig, John A Smolin (2014).

See also: enumerate_cliffords, clifford_cardinality.

QuantumClifford.tab — Method

Extract the underlying tableau structure.

julia> s = S"X"
+ X

julia> tab(s)
+ X

julia> tab(Destabilizer(s))
+ Z
+ X

julia> tab(MixedDestabilizer(s))
+ Z
+ X

julia> tab(CliffordOperator(s))
+ Z
+ X

julia> typeof(tab(CliffordOperator(s)))
Stabilizer{Vector{UInt8}, Matrix{UInt64}}

See also: stabilizerview, destabilizerview, logicalxview, logicalzview

QuantumClifford.tensor — Function

Tensor product between operators or tableaux. See also ⊗ and tensor_pow.

QuantumClifford.tensor_pow — Method

Repeated tensor product of an operators or a tableau. See also ⊗ and tensor_pow.

QuantumClifford.traceout! — Method

traceout!(s::Stabilizer, qubits; phases, rank) -> Any

Trace out a qubit.

QuantumClifford.traceout! — Method

traceout!(s::Union{MixedDestabilizer, MixedStabilizer}, qubits; phases, rank) -> Any

QuantumClifford.xbit — Method

Extract as a new bit array the X part of the UInt array of packed qubits of a given Pauli operator.

QuantumClifford.xview — Method

Get a view of the X part of the UInt array of packed qubits of a given Pauli operator.

QuantumClifford.zbit — Method

Extract as a new bit array the Z part of the UInt array of packed qubits of a given Pauli operator.

QuantumClifford.zview — Method

Get a view of the Y part of the UInt array of packed qubits of a given Pauli operator.