Full API

Autogenerated API list

QuantumSymbolics.StabilizerState — Type

State defined by a stabilizer tableau

julia> StabilizerState(S"XX ZZ")
𝒮₂

julia> express(StabilizerState(S"-X"))
Ket(dim=2)
  basis: Spin(1/2)
  0.7071067811865475 + 0.0im
 -0.7071067811865475 + 0.0im

QSymbolicsBase.CNOT — Constant

CNOT gate

QSymbolicsBase.CPHASE — Constant

CPHASE gate

QSymbolicsBase.Create — Constant

Creation operator, also available as the constant âꜛ - there is no unicode dagger superscript, so we use the uparrow

QSymbolicsBase.Destroy — Constant

Annihilation operator, also available as the constant â

QSymbolicsBase.F₁ — Constant

Single photon basis state of n

QSymbolicsBase.H — Constant

Hadamard gate

QSymbolicsBase.N — Constant

Number operator, also available as the constant n̂

QSymbolicsBase.Pm — Constant

Pauli "minus" operator, also available as the constant σ₋

QSymbolicsBase.Pp — Constant

Pauli "plus" operator, also available as the constant σ₊

QSymbolicsBase.X — Constant

Pauli X operator, also available as the constant σˣ

QSymbolicsBase.X1 — Constant

Basis state of σˣ

QSymbolicsBase.X2 — Constant

Basis state of σˣ

QSymbolicsBase.Y — Constant

Pauli Y operator, also available as the constant σʸ

QSymbolicsBase.Y1 — Constant

Basis state of σʸ

QSymbolicsBase.Y2 — Constant

Basis state of σʸ

QSymbolicsBase.Z — Constant

Pauli Z operator, also available as the constant σᶻ

QSymbolicsBase.Z1 — Constant

Basis state of σᶻ

QSymbolicsBase.Z2 — Constant

Basis state of σᶻ

QSymbolicsBase.vac — Constant

Vacuum basis state of n

QSymbolicsBase.CliffordRepr — Type

Representation using tableaux governed by QuantumClifford.jl

QSymbolicsBase.DephasingCPTP — Type

Single-qubit dephasing CPTP map

QSymbolicsBase.GateCPTP — Type

A unitary gate followed by a CPTP map

QSymbolicsBase.IdentityOp — Type

The identity operator for a given basis

julia> IdentityOp(X1⊗X2)
𝕀

julia> express(IdentityOp(Z2))
Operator(dim=2x2)
  basis: Spin(1/2)sparse([1, 2], [1, 2], ComplexF64[1.0 + 0.0im, 1.0 + 0.0im], 2, 2)

QSymbolicsBase.MixedState — Type

Completely depolarized state

julia> MixedState(X1⊗X2)
𝕄

julia> express(MixedState(X1⊗X2))
Operator(dim=4x4)
  basis: [Spin(1/2) ⊗ Spin(1/2)]sparse([1, 2, 3, 4], [1, 2, 3, 4], ComplexF64[0.25 + 0.0im, 0.25 + 0.0im, 0.25 + 0.0im, 0.25 + 0.0im], 4, 4)

  express(MixedState(X1⊗X2), CliffordRepr())
  Rank 0 stabilizer

  ━━━━
  + X_
  + _X
  ━━━━

  ━━━━
  + Z_
  + _Z

QSymbolicsBase.PauliNoiseCPTP — Type

Single-qubit Pauli noise CPTP map

julia> apply!(express(Z1), [1], PauliNoiseCPTP(1/4,1/4,1/4))
Operator(dim=2x2)
  basis: Spin(1/2)
 0.5+0.0im  0.0+0.0im
 0.0+0.0im  0.5+0.0im

QSymbolicsBase.QuantumMCRepr — Type

Similar to QuantumOpticsRepr, but using trajectories instead of superoperators.

QSymbolicsBase.QuantumOpticsRepr — Type

Representation using kets, densinty matrices, and superoperators governed by QuantumOptics.jl.

QSymbolicsBase.SProjector — Type

Projector for a given ket

julia> SProjector(X1⊗X2)
𝐏[|X₁⟩|X₂⟩]

julia> express(SProjector(X2))
Operator(dim=2x2)
  basis: Spin(1/2)
  0.5+0.0im  -0.5-0.0im
 -0.5+0.0im   0.5+0.0im