QuantumInterface.jl reference

QuantumInterface.jl is a base package meant to define common APIs used by many independent QIS packages.

Autogenerated API list for QuantumInterface

QuantumInterface.AbstractRepresentation — Type

An abstract type for the supported representation of quantum objects.

QuantumInterface.CliffordRepr — Type

Representation using tableaux governed by QuantumClifford.jl

QuantumInterface.QuantumOpticsRepr — Type

Representation using kets, bras, density matrices, and superoperators governed by QuantumOptics.jl.