Full API (autogenerated)

Unstable

This is experimental functionality with an unstable API.

QuantumClifford.Experimental.NoisyCircuits — Module

A module for simulating noisy Clifford circuits.

QuantumClifford.Experimental.NoisyCircuits.BellMeasurement — Type

A Bell measurement performing the correlation measurement corresponding to the given pauli projections on the qubits at the selected indices.

QuantumClifford.Experimental.NoisyCircuits.ConditionalGate — Type

A conditional gate that either performs truegate or falsegate, depending on the value of controlbit.

QuantumClifford.Experimental.NoisyCircuits.DecisionGate — Type

A conditional gate that performs one of the gates, depending on the output of decisionfunction applied to the entire classical bit register.

QuantumClifford.Experimental.NoisyCircuits.NoiseOp — Type

An operator that applies the given noise model to the qubits at the selected indices.

QuantumClifford.Experimental.NoisyCircuits.NoiseOpAll — Type

An operator that applies the given noise model to all qubits.

QuantumClifford.Experimental.NoisyCircuits.NoisyBellMeasurement — Type

A Bell measurement in which each of the measured qubits has a chance to have flipped.

QuantumClifford.Experimental.NoisyCircuits.NoisyGate — Type

A gate consisting of the given noise applied after the given perfect Clifford gate.

QuantumClifford.Experimental.NoisyCircuits.UnbiasedUncorrelatedNoise — Type

Depolarization noise model with total probability of error 3*errprobthird.

QuantumClifford.Experimental.NoisyCircuits.VerifyOp — Type

A "probe" to verify that the state of the qubits corresponds to a desired good_state, e.g. at the end of the execution of a circuit.

QuantumClifford.Experimental.NoisyCircuits.affectedqubits — Function

A method giving the qubits acted upon by a given operation. Part of the Noise interface.

QuantumClifford.Experimental.NoisyCircuits.applybranches — Function

Compute all possible new states after the application of the given operator. Reports the probability of each one of them. Deterministic, part of the Perturbative Expansion interface.

QuantumClifford.Experimental.NoisyCircuits.applynoise! — Function

A method modifying a given state by applying the corresponding noise model. Non-deterministic, part of the Noise interface.

QuantumClifford.Experimental.NoisyCircuits.applynoise_branches — Function

Compute all possible new states after the application of the given noise model. Reports the probability of each one of them. Deterministic, part of the Noise interface.

QuantumClifford.Experimental.NoisyCircuits.mctrajectories — Method

Run multiple Monte Carlo trajectories and report the aggregate final statuses of each.

QuantumClifford.Experimental.NoisyCircuits.mctrajectory! — Method

Run a single Monte Carlo sample, starting with (and modifying) state by applying the given circuit. Uses apply! under the hood.

QuantumClifford.Experimental.NoisyCircuits.petrajectories — Method

Run a perturbative expansion to a given order. This is the main public fuction for the perturbative expansion approach.

QuantumClifford.Experimental.NoisyCircuits.petrajectory — Method

Run a perturbative expansion to a given order. Uses applybranches under the hood.